Refactor OpenVDB Points Convert SOP to move more functionality into PointUtils
