+++ openvdb_houdini/houdini/SOP_OpenVDB_Points_Convert.cc
@@ -36,14 +36,12 @@
 
 #include <openvdb/openvdb.h>
 #include <openvdb/points/PointDataGrid.h>
-#include <openvdb/points/PointAttribute.h>
-#include <openvdb/points/PointConversion.h>
-#include <openvdb/points/PointGroup.h>
+#include <openvdb/points/PointCount.h>
+#include <openvdb/points/PointMask.h>
 
 #include <openvdb_houdini/Utils.h>
-#include <openvdb_houdini/SOP_NodeVDB.h>
 #include <openvdb_houdini/PointUtils.h>
-#include <openvdb_houdini/AttributeTransferUtil.h>
+#include <openvdb_houdini/SOP_NodeVDB.h>
 
 #include <houdini_utils/geometry.h>
 #include <houdini_utils/ParmFactory.h>
@@ -59,675 +57,16 @@
     #include <GU/GU_DetailHandle.h>
 #endif
 
-#include <CH/CH_Manager.h>
-#include <GA/GA_Types.h> // for GA_ATTRIB_POINT
-#include <SYS/SYS_Types.h> // for int32, float32, etc
+#include <CH/CH_Manager.h> // for CHgetEvalTime
 
-#include <algorithm>
-#include <map>
-#include <sstream>
-#include <stdexcept>
-#include <string>
-#include <type_traits>
-#include <utility> // for std::pair
-#include <vector>
 
 using namespace openvdb;
 using namespace openvdb::points;
 using namespace openvdb::math;
 
-namespace openvdb_houdini {
-    template <> inline openvdb::math::Quat<float>
-    evalAttrDefault<openvdb::math::Quat<float>>(const GA_Defaults& defaults, int)
-    {
-        openvdb::math::Quat<float> quat;
-        fpreal32 value;
-
-        for (int i = 0; i < 4; i++) {
-            defaults.get(i, value);
-            quat[i] = float(value);
-        }
-
-        return quat;
-    }
-
-    template <> inline openvdb::math::Quat<double>
-    evalAttrDefault<openvdb::math::Quat<double>>(const GA_Defaults& defaults, int)
-    {
-        openvdb::math::Quat<double> quat;
-        fpreal64 value;
-
-        for (int i = 0; i < 4; i++) {
-            defaults.get(i, value);
-            quat[i] = double(value);
-        }
-
-        return quat;
-    }
-
-    template <> inline openvdb::math::Mat4<float>
-    evalAttrDefault<openvdb::math::Mat4<float>>(const GA_Defaults& defaults, int)
-    {
-        openvdb::math::Mat4<float> mat;
-        fpreal64 value;
-        float* data = mat.asPointer();
-
-        for (int i = 0; i < 16; i++) {
-            defaults.get(i, value);
-            data[i] = float(value);
-        }
-
-        return mat;
-    }
-
-    template <> inline openvdb::math::Mat4<double>
-    evalAttrDefault<openvdb::math::Mat4<double>>(const GA_Defaults& defaults, int)
-    {
-        openvdb::math::Mat4<double> mat;
-        fpreal64 value;
-        double* data = mat.asPointer();
-
-        for (int i = 0; i < 16; i++) {
-            defaults.get(i, value);
-            data[i] = double(value);
-        }
-
-        return mat;
-    }
-}
 
 namespace hvdb = openvdb_houdini;
 namespace hutil = houdini_utils;
-
-enum CONVERSION_MODE
-{
-    MODE_CONVERT_TO_VDB = 0,
-    MODE_CONVERT_FROM_VDB,
-    MODE_GENERATE_MASK,
-    MODE_COUNT_POINTS,
-};
-
-enum COMPRESSION_TYPE
-{
-    NONE = 0,
-    TRUNCATE,
-    UNIT_VECTOR,
-    UNIT_FIXED_POINT_8,
-    UNIT_FIXED_POINT_16,
-};
-
-/// @brief Returns supported Storage types for conversion from GA_Attribute
-///
-inline GA_Storage
-attributeStorageType(const GA_Attribute* const attribute)
-{
-    if (!attribute) return GA_STORE_INVALID;
-
-    const GA_AIFTuple* tupleAIF = attribute->getAIFTuple();
-    if (!tupleAIF)
-    {
-        if (attribute->getAIFStringTuple())
-        {
-            return GA_STORE_STRING;
-        }
-    }
-    else
-    {
-        return tupleAIF->getStorage(attribute);
-    }
-
-    return GA_STORE_INVALID;
-}
-
-inline int16_t
-attributeTupleSize(const GA_Attribute* const attribute)
-{
-    if (!attribute) return int16_t(0);
-
-    const GA_AIFTuple* tupleAIF = attribute->getAIFTuple();
-    if (!tupleAIF)
-    {
-        const GA_AIFStringTuple* tupleAIFString = attribute->getAIFStringTuple();
-        if (tupleAIFString)
-        {
-            return static_cast<int16_t>(tupleAIFString->getTupleSize(attribute));
-        }
-    }
-    else
-    {
-        return static_cast<int16_t>(tupleAIF->getTupleSize(attribute));
-    }
-
-    return int16_t(0);
-}
-
-template <typename ValueType, typename CodecType = NullCodec>
-inline void
-convertAttributeFromHoudini(PointDataTree& tree, const tools::PointIndexTree& indexTree,
-    const openvdb::Name& name, const GA_Attribute* const attribute,
-    const GA_Defaults& defaults, const Index stride = 1)
-{
-    static_assert(!std::is_base_of<AttributeArray, ValueType>::value,
-        "ValueType must not be derived from AttributeArray");
-    static_assert(!std::is_same<ValueType, openvdb::Name>::value,
-        "ValueType must not be openvdb::Name/std::string");
-
-    using HoudiniAttribute = hvdb::HoudiniReadAttribute<ValueType>;
-
-    ValueType value = hvdb::evalAttrDefault<ValueType>(defaults, 0);
-
-    // empty metadata if default is zero
-    Metadata::Ptr defaultValue;
-    if (!math::isZero<ValueType>(value)) {
-        defaultValue = TypedMetadata<ValueType>(value).copy();
-    }
-
-    appendAttribute<ValueType, CodecType>(tree, name, zeroVal<ValueType>(),
-        stride, /*constantstride=*/true, defaultValue);
-
-    HoudiniAttribute houdiniAttribute(*attribute);
-    populateAttribute<PointDataTree, tools::PointIndexTree, HoudiniAttribute>(
-        tree, indexTree, name, houdiniAttribute, stride);
-}
-
-inline void
-convertAttributeFromHoudini(PointDataTree& tree, const tools::PointIndexTree& indexTree,
-    const openvdb::Name& name, const GA_Attribute* const attribute, const int compression = 0)
-{
-    using HoudiniStringAttribute = hvdb::HoudiniReadAttribute<openvdb::Name>;
-
-    if (!attribute) {
-        std::stringstream ss; ss << "Invalid attribute - " << attribute->getName();
-        throw std::runtime_error(ss.str());
-    }
-
-    const GA_Storage storage(attributeStorageType(attribute));
-
-    if (storage == GA_STORE_INVALID) {
-        std::stringstream ss; ss << "Invalid attribute type - " << attribute->getName();
-        throw std::runtime_error(ss.str());
-    }
-
-    const int16_t width(attributeTupleSize(attribute));
-    assert(width > 0);
-
-    // explicitly handle string attributes
-
-    if (storage == GA_STORE_STRING) {
-        appendAttribute<Name>(tree, name);
-        HoudiniStringAttribute houdiniAttribute(*attribute);
-        populateAttribute<PointDataTree, tools::PointIndexTree, HoudiniStringAttribute>(
-            tree, indexTree, name, houdiniAttribute);
-        return;
-    }
-
-    const GA_AIFTuple* tupleAIF = attribute->getAIFTuple();
-    if (!tupleAIF) {
-        std::stringstream ss; ss << "Invalid attribute type - " << attribute->getName();
-        throw std::runtime_error(ss.str());
-    }
-
-    GA_Defaults defaults = tupleAIF->getDefaults(attribute);
-    const GA_TypeInfo typeInfo(attribute->getOptions().typeInfo());
-
-    const bool isVector = width == 3 && (typeInfo == GA_TYPE_VECTOR ||
-                                         typeInfo == GA_TYPE_NORMAL ||
-                                         typeInfo == GA_TYPE_COLOR);
-    const bool isQuaternion = width == 4 && (typeInfo == GA_TYPE_QUATERNION);
-    const bool isMatrix = width == 16 && (typeInfo == GA_TYPE_TRANSFORM);
-
-    if (isVector)
-    {
-        if (storage == GA_STORE_INT32) {
-            convertAttributeFromHoudini<Vec3<int>>(tree, indexTree, name, attribute, defaults);
-        }
-        else if (storage == GA_STORE_REAL16)
-        {
-            // implicitly convert 16-bit float into truncated 32-bit float
-
-            convertAttributeFromHoudini<Vec3<float>, TruncateCodec>(
-                tree, indexTree, name, attribute, defaults);
-        }
-        else if (storage == GA_STORE_REAL32)
-        {
-            if (compression == NONE) {
-                convertAttributeFromHoudini<Vec3<float>>(
-                    tree, indexTree, name, attribute, defaults);
-            }
-            else if (compression == TRUNCATE) {
-                convertAttributeFromHoudini<Vec3<float>, TruncateCodec>(
-                    tree, indexTree, name, attribute, defaults);
-            }
-            else if (compression == UNIT_VECTOR) {
-                convertAttributeFromHoudini<Vec3<float>, UnitVecCodec>(
-                    tree, indexTree, name, attribute, defaults);
-            }
-            else if (compression == UNIT_FIXED_POINT_8) {
-                convertAttributeFromHoudini<Vec3<float>, FixedPointCodec<true, UnitRange>>(
-                    tree, indexTree, name, attribute, defaults);
-            }
-            else if (compression == UNIT_FIXED_POINT_16) {
-                convertAttributeFromHoudini<Vec3<float>, FixedPointCodec<false, UnitRange>>(
-                    tree, indexTree, name, attribute, defaults);
-            }
-        }
-        else if (storage == GA_STORE_REAL64) {
-            convertAttributeFromHoudini<Vec3<double>>(tree, indexTree, name, attribute, defaults);
-        }
-        else {
-            std::stringstream ss; ss << "Unknown vector attribute type - " << name;
-            throw std::runtime_error(ss.str());
-        }
-    }
-    else if (isQuaternion)
-    {
-        if (storage == GA_STORE_REAL16)
-        {
-            // implicitly convert 16-bit float into 32-bit float
-
-            convertAttributeFromHoudini<Quat<float>>(tree, indexTree, name, attribute, defaults);
-        }
-        else if (storage == GA_STORE_REAL32)
-        {
-            convertAttributeFromHoudini<Quat<float>>(tree, indexTree, name, attribute, defaults);
-        }
-        else if (storage == GA_STORE_REAL64) {
-            convertAttributeFromHoudini<Quat<double>>(tree, indexTree, name, attribute, defaults);
-        }
-        else {
-            std::stringstream ss; ss << "Unknown quaternion attribute type - " << name;
-            throw std::runtime_error(ss.str());
-        }
-    }
-    else if (isMatrix)
-    {
-        if (storage == GA_STORE_REAL16)
-        {
-            // implicitly convert 16-bit float into 32-bit float
-
-            convertAttributeFromHoudini<Mat4<float>>(tree, indexTree, name, attribute, defaults);
-        }
-        else if (storage == GA_STORE_REAL32)
-        {
-            convertAttributeFromHoudini<Mat4<float>>(tree, indexTree, name, attribute, defaults);
-        }
-        else if (storage == GA_STORE_REAL64) {
-            convertAttributeFromHoudini<Mat4<double>>(tree, indexTree, name, attribute, defaults);
-        }
-        else {
-            std::stringstream ss; ss << "Unknown matrix attribute type - " << name;
-            throw std::runtime_error(ss.str());
-        }
-    }
-    else {
-        if (storage == GA_STORE_BOOL) {
-            convertAttributeFromHoudini<bool>(tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_INT16) {
-            convertAttributeFromHoudini<int16_t>(tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_INT32) {
-            convertAttributeFromHoudini<int32_t>(tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_INT64) {
-            convertAttributeFromHoudini<int64_t>(tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_REAL16) {
-            convertAttributeFromHoudini<float, TruncateCodec>(
-                tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_REAL32 && compression == NONE) {
-            convertAttributeFromHoudini<float>(tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_REAL32 && compression == TRUNCATE) {
-            convertAttributeFromHoudini<float, TruncateCodec>(
-                tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_REAL32 && compression == UNIT_FIXED_POINT_8) {
-            convertAttributeFromHoudini<float, FixedPointCodec<true, UnitRange>>(
-                tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_REAL32 && compression == UNIT_FIXED_POINT_16) {
-            convertAttributeFromHoudini<float, FixedPointCodec<false, UnitRange>>(
-                tree, indexTree, name, attribute, defaults, width);
-        } else if (storage == GA_STORE_REAL64) {
-            convertAttributeFromHoudini<double>(tree, indexTree, name, attribute, defaults, width);
-        } else {
-            std::stringstream ss; ss << "Unknown attribute type - " << name;
-            throw std::runtime_error(ss.str());
-        }
-    }
-}
-
-
-////////////////////////////////////////
-
-
-using AttributeInfoMap = std::map<Name, std::pair<int, bool>>;
-
-
-///////////////////////////////////////
-
-
-inline
-PointDataGrid::Ptr
-createPointDataGrid(const GU_Detail& ptGeo, const int compression,
-                    const AttributeInfoMap& attributes, const openvdb::math::Transform& transform)
-{
-    using HoudiniPositionAttribute = hvdb::HoudiniReadAttribute<openvdb::Vec3d>;
-
-    // store point group information
-
-    const GA_ElementGroupTable& elementGroups = ptGeo.getElementGroupTable(GA_ATTRIB_POINT);
-
-    // Create PointPartitioner compatible P attribute wrapper (for now no offset filtering)
-
-    const GA_Attribute& positionAttribute = *ptGeo.getP();
-
-    hvdb::OffsetListPtr offsets;
-    hvdb::OffsetPairListPtr offsetPairs;
-
-    size_t vertexCount = 0;
-
-    for (GA_Iterator primitiveIt(ptGeo.getPrimitiveRange()); !primitiveIt.atEnd(); ++primitiveIt) {
-        const GA_Primitive* primitive = ptGeo.getPrimitiveList().get(*primitiveIt);
-
-        if (primitive->getTypeId() != GA_PRIMNURBCURVE) continue;
-
-        vertexCount = primitive->getVertexCount();
-
-        if (vertexCount == 0)  continue;
-
-        if (!offsets)   offsets.reset(new hvdb::OffsetList);
-
-        GA_Offset firstOffset = primitive->getPointOffset(0);
-        offsets->push_back(firstOffset);
-
-        if (vertexCount > 1) {
-            if (!offsetPairs)   offsetPairs.reset(new hvdb::OffsetPairList);
-
-            for (size_t i = 1; i < vertexCount; i++) {
-                GA_Offset offset = primitive->getPointOffset(i);
-                offsetPairs->push_back(hvdb::OffsetPair(firstOffset, offset));
-            }
-        }
-    }
-
-    HoudiniPositionAttribute points(positionAttribute, offsets);
-
-    // Create PointIndexGrid used for consistent index ordering in all attribute conversion
-
-    tools::PointIndexGrid::Ptr pointIndexGrid =
-        tools::createPointIndexGrid<tools::PointIndexGrid>(points, transform);
-
-    // Create PointDataGrid using position attribute
-
-    PointDataGrid::Ptr pointDataGrid;
-
-    if (compression == 1 /*FIXED_POSITION_16*/) {
-        pointDataGrid = createPointDataGrid<FixedPointCodec<false>, PointDataGrid>(
-            *pointIndexGrid, points, transform);
-    }
-    else if (compression == 2 /*FIXED_POSITION_8*/) {
-        pointDataGrid = createPointDataGrid<FixedPointCodec<true>, PointDataGrid>(
-            *pointIndexGrid, points, transform);
-    }
-    else /*NONE*/ {
-        pointDataGrid = createPointDataGrid<NullCodec, PointDataGrid>(
-            *pointIndexGrid, points, transform);
-    }
-
-    tools::PointIndexTree& indexTree = pointIndexGrid->tree();
-    PointDataTree& tree = pointDataGrid->tree();
-    PointDataTree::LeafIter leafIter = tree.beginLeaf();
-
-    if (!leafIter)  return pointDataGrid;
-
-    // Append (empty) groups to tree
-
-    std::vector<Name> groupNames;
-    groupNames.reserve(elementGroups.entries());
-
-    for (auto it = elementGroups.beginTraverse(), itEnd = elementGroups.endTraverse();
-        it != itEnd; ++it)
-    {
-        groupNames.push_back((*it)->getName().toStdString());
-    }
-
-    appendGroups(tree, groupNames);
-
-    // Set group membership in tree
-
-    const int64_t numPoints = ptGeo.getNumPoints();
-    std::vector<short> inGroup(numPoints, short(0));
-
-    for (auto it = elementGroups.beginTraverse(), itEnd = elementGroups.endTraverse();
-        it != itEnd; ++it)
-    {
-        // insert group offsets
-
-        GA_Offset start, end;
-        GA_Range range(**it);
-        for (GA_Iterator rangeIt = range.begin(); rangeIt.blockAdvance(start, end); ) {
-            end = std::min(end, numPoints);
-            for (GA_Offset off = start; off < end; ++off) {
-                assert(off < numPoints);
-                inGroup[off] = short(1);
-            }
-        }
-
-        const Name groupName = (*it)->getName().toStdString();
-        setGroup(tree, indexTree, inGroup, groupName);
-
-        std::fill(inGroup.begin(), inGroup.end(), short(0));
-    }
-
-    // Add other attributes to PointDataGrid
-
-    for (const auto& attrInfo : attributes)
-    {
-        const openvdb::Name& name = attrInfo.first;
-
-        // skip position as this has already been added
-
-        if (name == "P")  continue;
-
-        GA_ROAttributeRef attrRef = ptGeo.findPointAttribute(name.c_str());
-
-        if (!attrRef.isValid())     continue;
-
-        GA_Attribute const * gaAttribute = attrRef.getAttribute();
-
-        if (!gaAttribute)             continue;
-
-        const GA_AIFSharedStringTuple* sharedStringTupleAIF =
-            gaAttribute->getAIFSharedStringTuple();
-        const bool isString = bool(sharedStringTupleAIF);
-
-        // Extract all the string values from the string table and insert them
-        // into the Descriptor Metadata
-        if (isString)
-        {
-            // Iterate over the strings in the table and insert them into the Metadata
-            MetaMap& metadata = makeDescriptorUnique(tree)->getMetadata();
-            StringMetaInserter inserter(metadata);
-            for (auto it = sharedStringTupleAIF->begin(gaAttribute),
-                itEnd = sharedStringTupleAIF->end(); !(it == itEnd); ++it)
-            {
-                Name str(it.getString());
-                if (!str.empty())   inserter.insert(str);
-            }
-        }
-
-        convertAttributeFromHoudini(tree, indexTree, name, gaAttribute,
-            /*compression=*/attrInfo.second.first);
-    }
-
-    // Attempt to compact attributes
-
-    compactAttributes(tree);
-
-    // Apply blosc compression to attributes
-
-    for (const auto& attrInfo : attributes)
-    {
-        if (!attrInfo.second.second)  continue;
-
-        bloscCompressAttribute(tree, attrInfo.first);
-    }
-
-    return pointDataGrid;
-}
-
-///////////////////////////////////////
-
-
-template<typename ValueType>
-Metadata::Ptr
-createTypedMetadataFromAttribute(const GA_Attribute* const attribute, const uint32_t component = 0)
-{
-    using HoudiniAttribute = hvdb::HoudiniReadAttribute<ValueType>;
-
-    ValueType value;
-    HoudiniAttribute::get(*attribute, value, /*offset*/0, component);
-    return openvdb::TypedMetadata<ValueType>(value).copy();
-}
-
-template <typename ValueType>
-void
-populateHoudiniDetailAttribute(GA_RWAttributeRef& attrib, const openvdb::MetaMap& metaMap,
-                               const Name& key, const int index)
-{
-    using WriteHandleType = typename hvdb::GAHandleTraits<ValueType>::RW;
-    using TypedMetadataT = TypedMetadata<ValueType>;
-
-    typename TypedMetadataT::ConstPtr typedMetadata = metaMap.getMetadata<TypedMetadataT>(key);
-    if (!typedMetadata) return;
-
-    const ValueType& value = typedMetadata->value();
-    WriteHandleType handle(attrib.getAttribute());
-    hvdb::writeAttributeValue<WriteHandleType, ValueType>(handle, GA_Offset(0), index, value);
-}
-
-inline void
-convertGlobalMetadataToHoudini(GU_Detail& detail, const openvdb::MetaMap& metaMap,
-                               std::vector<std::string>& warnings)
-{
-    struct Local {
-        static bool isGlobalMetadata(const Name& name) {
-            return name.compare(0, 7, "global:") == 0;
-        }
-
-        static Name toDetailName(const Name& name) {
-            Name detailName(name);
-            detailName.erase(0, 7);
-            const size_t open = detailName.find('[');
-            if (open != std::string::npos) {
-                detailName = detailName.substr(0, open);
-            }
-            return detailName;
-        }
-
-        static int toDetailIndex(const Name& name) {
-            const size_t open = name.find('[');
-            const size_t close = name.find(']');
-            int index = 0;
-            if (open != std::string::npos && close != std::string::npos &&
-                close == name.length()-1 && open > 0 && open+1 < close) {
-                try { // parse array index
-                    index = std::stoi(name.substr(open+1, close-open-1));
-                }
-                catch (const std::exception&) {}
-            }
-            return index;
-        }
-    };
-
-    using DetailInfo = std::pair<Name, int>;
-    using DetailMap = std::map<Name, DetailInfo>;
-
-    DetailMap detailCreate;
-    DetailMap detailPopulate;
-
-    for(MetaMap::ConstMetaIterator iter = metaMap.beginMeta(); iter != metaMap.endMeta(); ++iter)
-    {
-        const Metadata::Ptr metadata = iter->second;
-        if (!metadata) continue;
-
-        const Name& key = iter->first;
-
-        if (!Local::isGlobalMetadata(key)) continue;
-
-        Name name = Local::toDetailName(key);
-        int index = Local::toDetailIndex(key);
-
-        // add to creation map
-
-        if (detailCreate.find(name) == detailCreate.end()) {
-            detailCreate[name] = DetailInfo(metadata->typeName(), index);
-        }
-        else {
-            if (index > detailCreate[name].second)   detailCreate[name].second = index;
-        }
-
-        // add to populate map
-
-        detailPopulate[key] = DetailInfo(name, index);
-    }
-
-    // add all detail attributes
-
-    for (const auto& item : detailCreate) {
-        const Name& name = item.first;
-        const DetailInfo& info = item.second;
-        const Name& type = info.first;
-        const int size = info.second;
-        GA_RWAttributeRef attribute = detail.findGlobalAttribute(name);
-
-        if (attribute.isInvalid())
-        {
-            const GA_Storage storage = hvdb::gaStorageFromAttrString(type);
-
-            if (storage == GA_STORE_INVALID) {
-                throw std::runtime_error("Invalid attribute storage type \"" + name + "\".");
-            }
-
-            if (type == "vec3s" || type == "vec3d") {
-                attribute = detail.addTuple(storage, GA_ATTRIB_GLOBAL, name.c_str(), 3);
-                attribute.setTypeInfo(GA_TYPE_VECTOR);
-            }
-            else {
-                attribute = detail.addTuple(storage, GA_ATTRIB_GLOBAL, name.c_str(), size+1);
-            }
-
-            if (!attribute.isValid()) {
-                throw std::runtime_error("Error creating attribute with name \"" + name + "\".");
-            }
-        }
-    }
-
-    // populate the values
-
-    for (const auto& item : detailPopulate) {
-        const Name& key = item.first;
-        const DetailInfo& info = item.second;
-        const Name& name = info.first;
-        const int index = info.second;
-        const Name& type = metaMap[key]->typeName();
-
-        GA_RWAttributeRef attrib = detail.findGlobalAttribute(name);
-        assert(!attrib.isInvalid());
-
-        if (type == openvdb::typeNameAsString<bool>())                 populateHoudiniDetailAttribute<bool>(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<int16_t>())         populateHoudiniDetailAttribute<int16_t>(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<int32_t>())         populateHoudiniDetailAttribute<int32_t>(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<int64_t>())         populateHoudiniDetailAttribute<int64_t>(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<float>())           populateHoudiniDetailAttribute<float>(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<double>())          populateHoudiniDetailAttribute<double>(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<Vec3<int32_t> >())  populateHoudiniDetailAttribute<Vec3<int32_t> >(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<Vec3<float> >())    populateHoudiniDetailAttribute<Vec3<float> >(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<Vec3<double> >())   populateHoudiniDetailAttribute<Vec3<double> >(attrib, metaMap, key, index);
-        else if (type == openvdb::typeNameAsString<Name>())            populateHoudiniDetailAttribute<Name>(attrib, metaMap, key, index);
-        else {
-            std::stringstream ss;
-            ss << "Metadata value \"" << key
-                << "\" unsupported type for detail attribute conversion.";
-            warnings.push_back(ss.str());
-        }
-    }
-}
 
 
 ////////////////////////////////////////
@@ -736,7 +75,20 @@ convertGlobalMetadataToHoudini(GU_Detail& detail, const openvdb::MetaMap& metaMa
 class SOP_OpenVDB_Points_Convert: public hvdb::SOP_NodeVDB
 {
 public:
-    enum { TRANSFORM_TARGET_POINTS = 0, TRANSFORM_VOXEL_SIZE, TRANSFORM_REF_GRID };
+    enum TRANSFORM_MODE
+    {
+        TRANSFORM_TARGET_POINTS = 0,
+        TRANSFORM_VOXEL_SIZE,
+        TRANSFORM_REF_GRID
+    };
+
+    enum CONVERSION_MODE
+    {
+        MODE_CONVERT_TO_VDB = 0,
+        MODE_CONVERT_FROM_VDB,
+        MODE_GENERATE_MASK,
+        MODE_COUNT_POINTS,
+    };
 
     SOP_OpenVDB_Points_Convert(OP_Network*, const char* name, OP_Operator*);
     ~SOP_OpenVDB_Points_Convert() override = default;
@@ -757,6 +109,7 @@ private:
 
 
 ////////////////////////////////////////
+
 
 namespace {
 
@@ -830,6 +183,7 @@ const PRM_ChoiceList PrimAttrMenu(
     PRM_ChoiceListType(PRM_CHOICELIST_REPLACE), sopBuildAttrMenu);
 
 } // unnamed namespace
+
 
 ////////////////////////////////////////
 
@@ -1275,7 +629,7 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
 
                 const MetaMap& metaMap = grid;
                 std::vector<std::string> warnings;
-                convertGlobalMetadataToHoudini(geo, metaMap, warnings);
+                hvdb::convertMetadataToHoudini(geo, metaMap, warnings);
                 if (warnings.size() > 0) {
                     for (const auto& warning: warnings) {
                         addWarning(SOP_MESSAGE, warning.c_str());
@@ -1383,14 +737,10 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
         }
 
         if (transformMode == TRANSFORM_TARGET_POINTS) {
-            using HoudiniPositionAttribute = hvdb::HoudiniReadAttribute<openvdb::Vec3R>;
-
             const int pointsPerVoxel = static_cast<int>(evalInt("pointspervoxel", 0, time));
-            HoudiniPositionAttribute positions(*(ptGeo->getP()));
-
             const float voxelSize =
-                openvdb::points::computeVoxelSize<HoudiniPositionAttribute, hvdb::Interrupter>(
-                    positions, pointsPerVoxel, matrix, /*rounding*/ 5, &mBoss);
+                hvdb::computeVoxelSizeFromHoudini(*ptGeo, pointsPerVoxel,
+                    matrix, /*rounding*/ 5, mBoss);
 
             matrix.preScale(Vec3d(voxelSize) / math::getScale(matrix));
             transform = Transform::createLinearTransform(matrix);
@@ -1401,7 +751,7 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
         }
 
         UT_String attrName;
-        AttributeInfoMap attributes;
+        openvdb_houdini::AttributeInfoMap attributes;
 
         GU_Detail nonConstDetail;
         const GU_Detail* detail;
@@ -1450,7 +800,7 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
 
                     if (!attribute) continue;
 
-                    const GA_Storage storage(attributeStorageType(attribute));
+                    const GA_Storage storage(hvdb::attributeStorageType(attribute));
 
                     // only tuple and string tuple attributes are supported
 
@@ -1459,7 +809,7 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
                         throw std::runtime_error(ss.str());
                     }
 
-                    const int16_t width(attributeTupleSize(attribute));
+                    const int16_t width(hvdb::attributeTupleSize(attribute));
                     assert(width > 0);
 
                     const GA_TypeInfo typeInfo(attribute->getOptions().typeInfo());
@@ -1476,31 +826,31 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
 
                     // check value compression compatibility with attribute type
 
-                    if (valueCompression != NONE) {
+                    if (valueCompression != hvdb::COMPRESSION_NONE) {
                         if (storage == GA_STORE_STRING) {
                             // disable value compression for strings and add a SOP warning
 
                             std::stringstream ss;
                             ss << "Value compression not supported on string attributes. "
                                 "Disabling compression for attribute \"" << attributeName << "\".";
-                            valueCompression = NONE;
+                            valueCompression = hvdb::COMPRESSION_NONE;
                             addWarning(SOP_MESSAGE, ss.str().c_str());
                         } else {
                             // disable value compression for incompatible types
                             // and add a SOP warning
 
-                            if (valueCompression == TRUNCATE &&
+                            if (valueCompression == hvdb::COMPRESSION_TRUNCATE &&
                                 (storage != GA_STORE_REAL32 || isQuaternion || isMatrix))
                             {
                                 std::stringstream ss;
                                 ss << "Truncate value compression only supported for 32-bit"
                                     " floating-point attributes. Disabling compression for"
                                     " attribute \"" << attributeName << "\".";
-                                valueCompression = NONE;
+                                valueCompression = hvdb::COMPRESSION_NONE;
                                 addWarning(SOP_MESSAGE, ss.str().c_str());
                             }
 
-                            if (valueCompression == UNIT_VECTOR &&
+                            if (valueCompression == hvdb::COMPRESSION_UNIT_VECTOR &&
                                 (storage != GA_STORE_REAL32 || !isVector))
                             {
                                 std::stringstream ss;
@@ -1508,12 +858,13 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
                                     " vector 3 x 32-bit floating-point attributes. "
                                     "Disabling compression for attribute \""
                                     << attributeName << "\".";
-                                valueCompression = NONE;
+                                valueCompression = hvdb::COMPRESSION_NONE;
                                 addWarning(SOP_MESSAGE, ss.str().c_str());
                             }
 
-                            const bool isUnit = (valueCompression == UNIT_FIXED_POINT_8
-                                || valueCompression == UNIT_FIXED_POINT_16);
+                            const bool isUnit =
+                                (valueCompression == hvdb::COMPRESSION_UNIT_FIXED_POINT_8
+                              || valueCompression == hvdb::COMPRESSION_UNIT_FIXED_POINT_16);
                             if (isUnit && (storage != GA_STORE_REAL32 || (width != 1 && !isVector)))
                             {
                                 std::stringstream ss;
@@ -1521,7 +872,7 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
                                     " 3 x 32-bit floating-point attributes. "
                                     "Disabling compression for attribute \""
                                     << attributeName << "\".";
-                                valueCompression = NONE;
+                                valueCompression = hvdb::COMPRESSION_NONE;
                                 addWarning(SOP_MESSAGE, ss.str().c_str());
                             }
                         }
@@ -1557,7 +908,7 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
 
                     if (!attribute) continue;
 
-                    const GA_Storage storage(attributeStorageType(attribute));
+                    const GA_Storage storage(hvdb::attributeStorageType(attribute));
 
                     // only tuple and string tuple attributes are supported
 
@@ -1566,7 +917,7 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
                         throw std::runtime_error(ss.str());
                     }
 
-                    const int16_t width(attributeTupleSize(attribute));
+                    const int16_t width(hvdb::attributeTupleSize(attribute));
                     assert(width > 0);
 
                     const GA_TypeInfo typeInfo(attribute->getOptions().typeInfo());
@@ -1574,16 +925,23 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
                     const bool isNormal = width == 3 && typeInfo == GA_TYPE_NORMAL;
                     const bool isColor = width == 3 && typeInfo == GA_TYPE_COLOR;
 
-                    int valueCompression = NONE;
+                    int valueCompression = hvdb::COMPRESSION_NONE;
 
                     if (isNormal) {
-                        if (normalCompression == 1)             valueCompression = UNIT_VECTOR;
-                        else if (normalCompression == 2)        valueCompression = TRUNCATE;
+                        if (normalCompression == 1) {
+                            valueCompression = hvdb::COMPRESSION_UNIT_VECTOR;
+                        } else if (normalCompression == 2) {
+                            valueCompression = hvdb::COMPRESSION_TRUNCATE;
+                        }
                     }
                     else if (isColor) {
-                        if (colorCompression == 1)              valueCompression = UNIT_FIXED_POINT_16;
-                        else if (colorCompression == 2)         valueCompression = UNIT_FIXED_POINT_8;
-                        else if (colorCompression == 3)         valueCompression = TRUNCATE;
+                        if (colorCompression == 1) {
+                            valueCompression = hvdb::COMPRESSION_UNIT_FIXED_POINT_16;
+                        } else if (colorCompression == 2) {
+                            valueCompression = hvdb::COMPRESSION_UNIT_FIXED_POINT_8;
+                        } else if (colorCompression == 3) {
+                            valueCompression = hvdb::COMPRESSION_TRUNCATE;
+                        }
                     }
 
                     // when converting all attributes apply no compression
@@ -1596,108 +954,14 @@ SOP_OpenVDB_Points_Convert::cookMySop(OP_Context& context)
 
         const int positionCompression = static_cast<int>(evalInt("poscompression", 0, time));
 
-        PointDataGrid::Ptr pointDataGrid = createPointDataGrid(
+        PointDataGrid::Ptr pointDataGrid = hvdb::convertHoudiniToPointDataGrid(
             *detail, positionCompression, attributes, *transform);
 
-        for (GA_AttributeDict::iterator iter = detail->attribs().begin(GA_SCOPE_PUBLIC);
-            !iter.atEnd(); ++iter)
-        {
-            const GA_Attribute* const attribute = *iter;
-            if (!attribute) continue;
+        std::vector<std::string> warnings;
+        hvdb::populateMetadataFromHoudini(*pointDataGrid, warnings, *detail);
 
-            const Name name("global:" + Name(attribute->getName()));
-            Metadata::Ptr metadata = (*pointDataGrid)[name];
-            if (metadata) continue;
-
-            const GA_Storage storage(attributeStorageType(attribute));
-            const int16_t width(attributeTupleSize(attribute));
-            const GA_TypeInfo typeInfo(attribute->getOptions().typeInfo());
-
-            const bool isVector = width == 3 && (typeInfo == GA_TYPE_VECTOR ||
-                                                 typeInfo == GA_TYPE_NORMAL ||
-                                                 typeInfo == GA_TYPE_COLOR);
-            const bool isQuaternion = width == 4 && (typeInfo == GA_TYPE_QUATERNION);
-            const bool isMatrix = width == 16 && (typeInfo == GA_TYPE_TRANSFORM);
-
-            if (isVector) {
-                if (storage == GA_STORE_REAL16) {
-                    metadata = createTypedMetadataFromAttribute<Vec3<float> >(attribute);
-                } else if (storage == GA_STORE_REAL32) {
-                    metadata = createTypedMetadataFromAttribute<Vec3<float> >(attribute);
-                } else if (storage == GA_STORE_REAL64) {
-                    metadata = createTypedMetadataFromAttribute<Vec3<double> >(attribute);
-                } else {
-                    std::stringstream ss;
-                    ss << "Detail attribute \"" << attribute->getName() << "\" " <<
-                        "unsupported vector type for metadata conversion.";
-                    addWarning(SOP_MESSAGE, ss.str().c_str());
-                    continue;
-                }
-                assert(metadata);
-                pointDataGrid->insertMeta(name, *metadata);
-            } else if (isQuaternion) {
-                if (storage == GA_STORE_REAL16) {
-                    metadata = createTypedMetadataFromAttribute<Quat<float>>(attribute);
-                } else if (storage == GA_STORE_REAL32) {
-                    metadata = createTypedMetadataFromAttribute<Quat<float>>(attribute);
-                } else if (storage == GA_STORE_REAL64) {
-                    metadata = createTypedMetadataFromAttribute<Quat<double>>(attribute);
-                } else {
-                    std::stringstream ss;
-                    ss << "Detail attribute \"" << attribute->getName() << "\" " <<
-                        "unsupported quaternion type for metadata conversion.";
-                    addWarning(SOP_MESSAGE, ss.str().c_str());
-                    continue;
-                }
-            } else if (isMatrix) {
-                if (storage == GA_STORE_REAL16) {
-                    metadata = createTypedMetadataFromAttribute<Mat4<float>>(attribute);
-                } else if (storage == GA_STORE_REAL32) {
-                    metadata = createTypedMetadataFromAttribute<Mat4<float>>(attribute);
-                } else if (storage == GA_STORE_REAL64) {
-                    metadata = createTypedMetadataFromAttribute<Mat4<double>>(attribute);
-                } else {
-                    std::stringstream ss;
-                    ss << "Detail attribute \"" << attribute->getName() << "\" " <<
-                        "unsupported matrix type for metadata conversion.";
-                    addWarning(SOP_MESSAGE, ss.str().c_str());
-                    continue;
-                }
-            } else {
-                for (int i = 0; i < width; i++) {
-                    if (storage == GA_STORE_BOOL) {
-                        metadata = createTypedMetadataFromAttribute<bool>(attribute, i);
-                    } else if (storage == GA_STORE_INT16) {
-                        metadata = createTypedMetadataFromAttribute<int16_t>(attribute, i);
-                    } else if (storage == GA_STORE_INT32) {
-                        metadata = createTypedMetadataFromAttribute<int32_t>(attribute, i);
-                    } else if (storage == GA_STORE_INT64) {
-                        metadata = createTypedMetadataFromAttribute<int64_t>(attribute, i);
-                    } else if (storage == GA_STORE_REAL16) {
-                        metadata = createTypedMetadataFromAttribute<float>(attribute, i);
-                    } else if (storage == GA_STORE_REAL32) {
-                        metadata = createTypedMetadataFromAttribute<float>(attribute, i);
-                    } else if (storage == GA_STORE_REAL64) {
-                        metadata = createTypedMetadataFromAttribute<double>(attribute, i);
-                    } else if (storage == GA_STORE_STRING) {
-                        metadata = createTypedMetadataFromAttribute<openvdb::Name>(attribute, i);
-                    } else {
-                        std::stringstream ss;
-                        ss << "Detail attribute \"" << attribute->getName() << "\" " <<
-                            "unsupported type for metadata conversion.";
-                        addWarning(SOP_MESSAGE, ss.str().c_str());
-                        continue;
-                    }
-                    assert(metadata);
-                    if (width > 1) {
-                        const Name arrayName(name + Name("[") + std::to_string(i) + Name("]"));
-                        pointDataGrid->insertMeta(arrayName, *metadata);
-                    }
-                    else {
-                        pointDataGrid->insertMeta(name, *metadata);
-                    }
-                }
-            }
+        for (const auto& warning : warnings) {
+            addWarning(SOP_MESSAGE, warning.c_str());
         }
 
         UT_String nameStr = "";
